完善套餐管理业务逻辑

+++ sky-server/src/main/java/com/sky/service/impl/SetmealServiceImpl.java
@@ -9,8 +9,11 @@ import com.sky.dto.SetmealDTO;
 import com.sky.dto.SetmealPageQueryDTO;
 import com.sky.entity.Setmeal;
 import com.sky.entity.SetmealDish;
+import com.sky.exception.BaseException;
 import com.sky.exception.DeletionNotAllowedException;
+import com.sky.exception.SetmealEnableFailedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
 import com.sky.mapper.SetmealDishMapper;
 import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
@@ -23,6 +26,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.management.InstanceAlreadyExistsException;
 import java.util.List;
 import java.util.Set;
 
@@ -34,6 +38,8 @@ public class SetmealServiceImpl implements SetmealService {
     private SetmealDishMapper setmealDishMapper;
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
     @Override
     public PageResult pageQuery(SetmealPageQueryDTO setmealPageQueryDTO) {
         PageHelper.startPage(setmealPageQueryDTO.getPage(),setmealPageQueryDTO.getPageSize());
@@ -44,18 +50,31 @@ public class SetmealServiceImpl implements SetmealService {
     @Override
     @Transactional
     public void add(SetmealDTO setmealDTO) {
+        // 套餐名是否重复
+        Setmeal setmeal = setmealMapper.getByName(setmealDTO.getName());
+        if(setmeal != null){
+            throw new BaseException(MessageConstant.SETMEAL_NAME_ALREADY_EXISTS);
+        }
+        // 分类为null 或不存在？
+        if(setmeal.getCategoryId() == null || categoryMapper.getNameById(setmeal.getCategoryId()) == null){
+            throw new BaseException(MessageConstant.SETMEAL_CATEGORY_ERROR);
+        }
+        List<SetmealDish> setmealDishes = setmealDTO.getSetmealDishes();
+        // 套餐无菜品？
+        if(setmealDishes == null || setmealDishes.isEmpty()){
+            throw new BaseException(MessageConstant.SETMEAL_DISHES_EMPTY);
+        }
         // 向setmeal插入一条数据
-        Setmeal setmeal = new Setmeal();
+        setmeal = new Setmeal();
         BeanUtils.copyProperties(setmealDTO, setmeal);
+        // 新增套餐默认停售状态
+        setmeal.setStatus(StatusConstant.DISABLE);
         setmealMapper.insert(setmeal);
         // 拿套餐id
         Long setmealId = setmeal.getId();
         // 向setmeal_dish批量插入菜品数据
-        List<SetmealDish> setmealDishes = setmealDTO.getSetmealDishes();
-        if(setmealDishes != null && !setmealDishes.isEmpty()){
-            setmealDishes.forEach(setmealDish -> setmealDish.setSetmealId(setmealId));
-            setmealDishMapper.insertBatch(setmealDishes);
-        }
+        setmealDishes.forEach(setmealDish -> setmealDish.setSetmealId(setmealId));
+        setmealDishMapper.insertBatch(setmealDishes);
     }
 
     @Override
@@ -74,17 +93,29 @@ public class SetmealServiceImpl implements SetmealService {
 
     @Override
     public void edit(SetmealDTO setmealDTO) {
+        // 套餐名是否重复
+        Setmeal setmeal = setmealMapper.getByName(setmealDTO.getName());
+        if(setmeal != null && setmeal.getId() != setmealDTO.getId()){
+            throw new BaseException(MessageConstant.SETMEAL_NAME_ALREADY_EXISTS);
+        }
+        // 分类为null 或不存在？
+        if(setmeal.getCategoryId() == null || categoryMapper.getNameById(setmeal.getCategoryId()) == null){
+            throw new BaseException(MessageConstant.SETMEAL_CATEGORY_ERROR);
+        }
+        List<SetmealDish> setmealDishes = setmealDTO.getSetmealDishes();
+        // 套餐无菜品？
+        if(setmealDishes == null || setmealDishes.isEmpty()){
+            throw new BaseException(MessageConstant.SETMEAL_DISHES_EMPTY);
+        }
+
         // 更新 setmeal 表
-        Setmeal setmeal = new Setmeal();
+        setmeal = new Setmeal();
         BeanUtils.copyProperties(setmealDTO, setmeal);
         setmealMapper.update(setmeal);
         // 更新 setmeal_dish 表
         setmealDishMapper.deleteDishes(setmeal.getId());
-        List<SetmealDish> setmealDishes = setmealDTO.getSetmealDishes();
-        if(setmealDishes != null && !setmealDishes.isEmpty()){
-            setmealDishes.forEach(setmealDish -> setmealDish.setSetmealId(setmealDTO.getId()));
-            setmealDishMapper.insertBatch(setmealDishes);
-        }
+        setmealDishes.forEach(setmealDish -> setmealDish.setSetmealId(setmealDTO.getId()));
+        setmealDishMapper.insertBatch(setmealDishes);
     }
 
     @Override
@@ -107,6 +138,16 @@ public class SetmealServiceImpl implements SetmealService {
     @Override
     public void changeStatus(int status, Long id) {
         Setmeal setmeal = setmealMapper.getById(id);
+        // 若套餐启售，但包含未启售菜品，不能启售
+        if (status == StatusConstant.ENABLE){
+            List<SetmealDish> setmealDishes = setmealDishMapper.getDishesBySetmealId(id);
+            for(SetmealDish setmealDish : setmealDishes){
+                Dish dish = dishMapper.getById(setmealDish.getDishId());
+                if(dish.getStatus() == StatusConstant.DISABLE){
+                    throw new SetmealEnableFailedException(MessageConstant.SETMEAL_ENABLE_FAILED);
+                }
+            }
+        }
         setmeal.setStatus(status);
         setmealMapper.update(setmeal);
     }
